+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Version History
 ===============
 
-1.0.0b2 2019-01-22
+1.0.0b2 2019-01-31
 ------------------
 
 `GitHub milestone <https://github.com/pika/pika/milestone/8>`_
@@ -10,6 +10,9 @@ Version History
 - ``BlockingConnection.consume`` now returns ``(None, None, None)`` when inactivity timeout is reached (`PR <https://github.com/pika/pika/pull/899>`_)
 - Python 3.7 support (`Issue <https://github.com/pika/pika/issues/1107>`_)
 - ``all_channels`` parameter of the ``Channel.basic_qos`` method renamed to ``global_qos``
+ - The `backpressure_detection` option of `ConnectionParameters` and
+   `URLParameters` property is REMOVED in favor of `Connection.Blocked` and
+   `Connection.Unblocked`. See `Connection.add_on_connection_blocked_callback`.
 
 
 **IMPORTANT**: The signature of the following methods has changed from Pika 0.13.0. In general, the callback parameter that indicates completion of the method has been moved to the end of the parameter list to be consistent with other parts of Pika's API and with other libraries in general.